# optimized histogram