@Cyan4973 2020-07-27
Interestingly, I've been trying to reproduce the result on my desktop, using a slightly different CPU (9700k), a different compiler by default (`gcc-10`) and a different OS (Linux, Ubuntu 20.04).

And I don't reproduce the speed gains of `histogram` function in micro-benchmark.

Interestingly, once combined into a larger `fse_compress.c` workload, it does produce some speed benefit.
It could be either an indirect benefit from using less memory for instructions, or once again a random impact from instruction alignment.